Update README: --num-threads
Was `--alignment-threads`. This is the same as `-j`,
and it is provided by **pbcopper**.

* https://github.com/PacificBiosciences/pbbioconda/issues/407

Note that the README in our GH repo,
    https://github.com/PacificBiosciences/pbmm2/
does not necessarily match what is released in Bioconda. And Bioconda
is actually several releases old now (1.4.0 from September 2020).
So we need to push to GH for this update, but we also need a
need Bioconda release in general.

TAK-1547